requirement.txt
pandas
matplotlib 
seaborn 
fpdf2 
numpy